Allow posts to be updated to remove featured status on save
Once a post is updated to be featured, it is not possible to uncheck
the featured checkbox and save it to remove featured status since the
isset post featured variable check can never be satisified.  Added
some javascript to append a hidden input element to the form when the
value is unchecked and remove the hidden input element from the form
when the value is checked.